Sherlock and Permutations EC: EC#4
This code asks the user to specify how many ones and zeros they want. The code then calculates the total number of unique arrangements with the specified amount of zeros and ones, while also finding the number of arrangements that start with 1.

Note that the code doesn't write out every single combination, and has two separate methods to finding both numbers. One is more logical and goes about solving the problem with a logical approach, however it has a tremendously exponential cost. The second method uses a mathematical formula to find both numbers, which saves a lot of time.